+++ src/board/setup-board.js
@@ -1,0 +1,7 @@
+import createBoardArray from './create-board.js';
+
+export default function setupBoard() {
+	createBoardArray('rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1');
+	global.moveList = [];
+	global.moveList.push('rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1');
+}
