+++ src/board/undo.js
@@ -1,7 +1,8 @@
-import createBoardArray from "./create-board.js";
+import createBoardArray from './create-board.js';
 
 export default function undoMove() {
 	global.moveList.pop();
-	createBoardArray(global.moveList.slice(-1)[0]);
-	return global.moveList.slice(-1)[0];
+	let lastMove = global.moveList.slice(-1)[0];
+	createBoardArray(lastMove);
+	return lastMove;
 }
