+++ src/helpers.js
@@ -1,11 +1,3 @@
-import createBoardArray from './board/create-board.js';
-
-export function setupBoard() {
-	createBoardArray('rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1');
-	global.moveList = [];
-	global.moveList.push('rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1');
-}
-
 export function indexToLetter(n) {
 	// 1-indexed; 0x40=uppercase, 0x60=lowercase
 	return String.fromCharCode(n + 0x40);
